Experiment with better support for implicit arguments
Allow implicit arguments to be marked in valspecs, like so:

val Zeros : forall 'n, 'n >= 0. implicit('n) unit -> bits('n)

The AST has been modified with an extra argument on Typ_fn to make
this work.

This is essentially the same as (atom('n), unit) -> bits('n), except
the first argument can be omitted, so the function can either be
called as in:

function f ('n : int) -> unit = {
  let x : bits(64) = Zeros();
  print_bits("x = ", x);
}

or

function f ('n : int) -> unit = {
  let x : bits(64) = Zeros(64, ());
  print_bits("x = ", x);
}

In the first case the first argument is filled in by the typechecker,
which when combined with sizeof-rewriting in the typechecker should
remove the need for a separate sizeof-rewriting pass in the rewriter.